requirements.txt.txt
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
attrs==24.3.0
backoff==2.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
filelock==3.13.1
filetype==1.2.0
filterpy==1.4.5
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
idna==3.7
imageio==2.36.1
inference-sdk==0.32.0
Jinja2==3.1.3
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==10.2.0
propcache==0.2.1
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.50
scikit-image==0.25.0
scipy==1.14.1
seaborn==0.13.2
setuptools==70.0.0
six==1.17.0
supervision
sympy==1.13.1
thop==0.1.1.post2209072238
tifffile==2024.12.12
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.2
ultralytics==8.2.22
ultralytics-thop==2.0.13
urllib3==2.3.0
yarl==1.18.3
